Update new payment

+++ services/basket/src/main/java/dev/berke/app/basket/BasketConfirmRequest.java
@@ -1,6 +1,6 @@
 package dev.berke.app.basket;
 
 public record BasketConfirmRequest(
-                Integer productId,
-                Double quantity) {
+        Integer productId,
+        Integer quantity) {
 }
