+++ services/basket/src/main/java/dev/berke/app/basket/BasketItemRequest.java
@@ -1,6 +1,6 @@
 package dev.berke.app.basket;
 
 public record BasketItemRequest(
-                Integer productId,
-                Double quantity) {
+        Integer productId,
+        Integer quantity) {
 }
