+++ services/basket/src/main/java/dev/berke/app/basket/ProductResponse.java
@@ -1,10 +1,11 @@
 package dev.berke.app.basket;
 
+import java.math.BigDecimal;
+
 public record ProductResponse(
-        Integer id,
-        String name,
-        Integer categoryId,
-        ItemType itemType,
-        Double price
-) {
+                Integer id,
+                String name,
+                Integer categoryId,
+                ItemType itemType,
+                BigDecimal price) {
 }
