+++ services/notification/src/main/java/dev/berke/app/kafka/order/Product.java
@@ -4,9 +4,9 @@ import java.math.BigDecimal;
 
 public record Product(
 
-                Integer productId,
-                String name,
-                String description,
-                BigDecimal price,
-                Double quantity) {
+        Integer productId,
+        String name,
+        String description,
+        BigDecimal price,
+        Integer quantity) {
 }
