+++ services/order/src/main/java/dev/berke/app/kafka/OrderConfirmation.java
@@ -4,13 +4,13 @@ import dev.berke.app.basket.BasketItem;
 import dev.berke.app.customer.CustomerResponse;
 import dev.berke.app.order.PaymentMethod;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record OrderConfirmation(
-        String orderReference,
-        Double totalAmount,
-        PaymentMethod paymentMethod,
-        CustomerResponse customer,
-        List<BasketItem> basketItems
-) {
+                String orderReference,
+                BigDecimal totalAmount,
+                PaymentMethod paymentMethod,
+                CustomerResponse customer,
+                List<BasketItem> basketItems) {
 }
