+++ services/order/src/main/java/dev/berke/app/orderline/OrderLineRequest.java
@@ -1,10 +1,8 @@
 package dev.berke.app.orderline;
 
 public record OrderLineRequest(
-
-        Integer id,
-        Integer orderId,
-        Integer productId,
-        double quantity
-) {
+                Integer id,
+                Integer orderId,
+                Integer productId,
+                Integer quantity) {
 }
