+++ services/order/src/main/java/dev/berke/app/product/PurchaseRequest.java
@@ -5,10 +5,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
 public record PurchaseRequest(
-        @NotNull(message = OrderConstants.PRODUCT_ID_NOT_NULL_MESSAGE)
-        Integer productId,
+                @NotNull(message = OrderConstants.PRODUCT_ID_NOT_NULL_MESSAGE) Integer productId,
 
-        @Positive(message = OrderConstants.QUANTITY_POSITIVE_MESSAGE)
-        double quantity
-) {
+                @Positive(message = OrderConstants.QUANTITY_POSITIVE_MESSAGE) Integer quantity) {
 }
