+++ services/order/src/main/java/dev/berke/app/product/PurchaseResponse.java
@@ -4,10 +4,9 @@ import java.math.BigDecimal;
 
 public record PurchaseResponse(
 
-        Integer productId,
-        String name,
-        String description,
-        BigDecimal price,
-        double quantity
-) {
+                Integer productId,
+                String name,
+                String description,
+                BigDecimal price,
+                Integer quantity) {
 }
