+++ services/product/src/main/java/dev/berke/app/product/ProductMapper.java
@@ -6,43 +6,39 @@ import org.springframework.stereotype.Service;
 @Service
 public class ProductMapper {
 
-    public Product toProduct(ProductRequest productRequest) {
-        return Product.builder()
-                .id(productRequest.id())
-                .name(productRequest.name())
-                .description(productRequest.description())
-                .availableQuantity(productRequest.availableQuantity())
-                .price(productRequest.price())
-                .category(
-                        Category.builder()
-                                .id(productRequest.categoryId())
-                                .build()
-                )
-                .build();
-    }
+        public Product toProduct(ProductRequest productRequest) {
+                return Product.builder()
+                                .id(productRequest.id())
+                                .name(productRequest.name())
+                                .description(productRequest.description())
+                                .availableQuantity(productRequest.availableQuantity())
+                                .price(productRequest.price())
+                                .category(
+                                                Category.builder()
+                                                                .id(productRequest.categoryId())
+                                                                .build())
+                                .build();
+        }
 
-    public ProductPurchaseResponse toProductPurchaseResponse(
-            Product product, double quantity
-    ) {
-        return new ProductPurchaseResponse(
-                product.getId(),
-                product.getName(),
-                product.getDescription(),
-                product.getPrice(),
-                quantity
-        );
-    }
+        public ProductPurchaseResponse toProductPurchaseResponse(
+                        Product product, Integer quantity) {
+                return new ProductPurchaseResponse(
+                                product.getId(),
+                                product.getName(),
+                                product.getDescription(),
+                                product.getPrice(),
+                                quantity);
+        }
 
-    public ProductResponse toProductResponse(Product product) {
-        return new ProductResponse(
-                product.getId(),
-                product.getName(),
-                product.getDescription(),
-                product.getAvailableQuantity(),
-                product.getPrice(),
-                product.getCategory().getId(),
-                product.getCategory().getName(),
-                product.getCategory().getDescription()
-        );
-    }
+        public ProductResponse toProductResponse(Product product) {
+                return new ProductResponse(
+                                product.getId(),
+                                product.getName(),
+                                product.getDescription(),
+                                product.getAvailableQuantity(),
+                                product.getPrice(),
+                                product.getCategory().getId(),
+                                product.getCategory().getName(),
+                                product.getCategory().getDescription());
+        }
 }
