+++ services/product/src/main/java/dev/berke/app/product/ProductPurchaseRequest.java
@@ -5,10 +5,7 @@ import jakarta.validation.constraints.NotNull;
 
 public record ProductPurchaseRequest(
 
-        @NotNull(message = ProductConstants.PRODUCT_ID_MANDATORY_MESSAGE)
-        Integer productId,
+                @NotNull(message = ProductConstants.PRODUCT_ID_MANDATORY_MESSAGE) Integer productId,
 
-        @NotNull(message = ProductConstants.QUANTITY_MANDATORY_MESSAGE)
-        double quantity
-) {
+                @NotNull(message = ProductConstants.QUANTITY_MANDATORY_MESSAGE) Integer quantity) {
 }
